Simplify and improve the simple sample Kotlin app
