+++ sample-kotlin/src/main/AndroidManifest.xml
@@ -3,19 +3,17 @@
   xmlns:tools="http://schemas.android.com/tools"
   package="net.helpscout.samples.beacon.kotlin">
 
-  <!-- appComponentFactory had to be defined as there is a sister project that uses AndroidX -->
   <application
-    android:name=".MainApplication"
+    android:name=".SampleApplication"
     android:allowBackup="true"
-    android:appComponentFactory="android.support.v4.app.CoreComponentFactory"
     android:icon="@mipmap/ic_launcher"
     android:label="@string/app_name"
     android:networkSecurityConfig="@xml/debug_trust_user_certs"
     android:roundIcon="@mipmap/ic_launcher_round"
-    android:supportsRtl="true"
+    android:supportsRtl="false"
     android:theme="@style/AppTheme"
-    tools:ignore="GoogleAppIndexingWarning"
-    tools:replace="android:appComponentFactory, android:supportsRtl">
+    tools:ignore="GoogleAppIndexingWarning,UnusedAttribute"
+    tools:replace="android:supportsRtl">
 
     <activity
       android:name="net.helpscout.samples.beacon.kotlin.MainActivity"
@@ -26,6 +24,7 @@
         <category android:name="android.intent.category.LAUNCHER" />
       </intent-filter>
     </activity>
+
   </application>
 
 </manifest>
